update ut for paw
