+++ source/module_cell/module_paw/test/rhoij1.dat
@@ -1,72 +1,190 @@
- 0.02188703434
- -0.06673644896
- 0.2034882207
- -1.134543361e-07
- 2.710666919e-06
- 0.06620713599
- 5.86463651e-08
- 3.894111841e-07
- 1.947941626e-06
- 0.06621693969
- -6.468749232e-07
- 1.51923826e-06
- 1.255978595e-06
- 5.153249335e-07
- 0.06620951375
- 1.53729588e-07
- -3.487439441e-06
- -0.08452117497
- -2.227129643e-06
- -1.369382834e-06
- 0.1079011939
- -7.26293686e-08
- -5.039140166e-07
- -2.49096291e-06
- -0.08453259842
- -1.276329531e-06
- 2.848537829e-06
- 0.1079143831
- 8.368731635e-07
- -1.973194344e-06
- -8.392804473e-07
- -1.292468003e-06
- -0.08452377862
- 7.72681564e-07
- 2.439506106e-06
- 0.1079039665
- 0.02188709511
- -0.06673764473
- 0.2034949456
- -4.101907869e-07
- 2.642518262e-06
- 0.06620987251
- 3.191326334e-07
- 1.813362146e-07
- -1.817637982e-06
- 0.06620277977
- -8.614672357e-08
- 6.148742837e-07
- 6.523376467e-07
- -2.906022646e-06
- 0.06620872924
- 5.48607772e-07
- -3.449561915e-06
- -0.08452473818
- 2.564476759e-06
- -7.31685773e-07
- 0.1079058318
- -4.045190771e-07
- -2.403128085e-07
- 2.008679449e-06
- -0.0845152669
- 3.157923971e-06
- -2.875868186e-06
- 0.1078932088
- 1.090596776e-07
- -7.821587661e-07
- -6.250318133e-08
- 2.668230728e-06
- -0.08452280732
- -4.928130275e-08
- -2.701677503e-06
- 0.1079027651
+30
+1
+2
+3
+4
+5
+6
+7
+8
+9
+10
+11
+12
+13
+14
+15
+16
+17
+18
+20
+21
+25
+26
+28
+29
+30
+31
+32
+33
+34
+36
+1.723931059770e+00
+1.604764521716e-02
+1.494012537960e-04
+-1.027645132703e-03
+1.199346602637e-04
+1.788384419289e+00
+-4.104145515715e-10
+1.541392006352e-10
+-1.028153466025e-08
+1.794463509848e+00
+-4.463068544051e-03
+-1.640962508172e-04
+-4.230665971256e-03
+-4.097241934977e-09
+1.791680468949e+00
+-1.019548737813e-04
+3.952629922034e-06
+6.766369380865e-02
+-1.896852009408e-04
+2.560065063630e-03
+6.812757305568e-02
+1.930432934384e-10
+2.586492388943e-03
+-3.616881261683e-04
+-8.057331932562e-06
+-6.603249854859e-04
+1.731662673311e-10
+6.804684119605e-02
+-2.610124355388e-05
+2.584535182494e-03
+
+15
+1
+2
+3
+4
+5
+6
+7
+8
+9
+10
+11
+12
+13
+14
+15
+3.084946019717e+01
+-1.513653458902e+01
+7.431394067537e+00
+2.965701410228e+00
+-1.461691126642e+00
+3.026339681046e-01
+1.277491351392e-08
+-6.266671342595e-09
+1.087802717879e-09
+8.414979146768e-02
+-1.057957404592e+00
+5.201186935953e-01
+-7.874127927236e-02
+-1.550442386653e-09
+1.105231926177e-01
+
+15
+1
+2
+3
+4
+5
+6
+7
+8
+9
+10
+11
+12
+13
+14
+15
+3.072639483959e+01
+-1.506348599946e+01
+7.389507590546e+00
+-1.478257306558e+00
+7.287490806885e-01
+1.374847941453e-01
+-2.527777445980e+00
+1.243485050039e+00
+9.346004321031e-02
+2.430931259861e-01
+-1.035282171321e+00
+5.118985921631e-01
+3.745607317851e-02
+6.530419132712e-02
+1.102101358232e-01
+
+15
+1
+2
+3
+4
+5
+6
+7
+8
+9
+10
+11
+12
+13
+14
+15
+3.072639499832e+01
+-1.506348617326e+01
+7.389507722114e+00
+-1.478257326751e+00
+7.287490947226e-01
+1.374847919726e-01
+2.527777495043e+00
+-1.243485081310e+00
+-9.346004816168e-02
+2.430931307570e-01
+-1.035282200554e+00
+5.118986094995e-01
+3.745607423306e-02
+-6.530419621766e-02
+1.102101347655e-01
+
+15
+1
+2
+3
+4
+5
+6
+7
+8
+9
+10
+11
+12
+13
+14
+15
+3.100425217119e+01
+-1.522680161652e+01
+7.482624648216e+00
+6.360854447087e-03
+-3.831398134964e-03
+8.777270655541e-02
+6.816615280579e-09
+-3.163414023143e-09
+-1.483636330591e-09
+8.401000380891e-02
+3.149626535913e+00
+-1.553135270239e+00
+-2.375993453642e-04
+5.330824319185e-10
+3.289138736057e-01
+
