+++ source/module_cell/module_paw/test/test_paw1.cpp
@@ -198,259 +198,3 @@ TEST_F(Test_Leg_Pol, test_paw)
         EXPECT_NEAR(ref,result,1e-8);
     }
 }
-
-class Test_PAW_Cell_k : public testing::Test
-{
-    protected:
-
-    Paw_Cell paw_cell;
-
-    double ecut = 20.0, cell_factor = 1.2, omega = 265.302, tpiba = 0.6159985595;
-    int nat = 2, ntyp = 1;
-    int atom_type[2] = {0,0}; // Si, Si
-    std::vector<std::string> filename_list;
-    int nx = 24, ny = 24, nz = 24;
-    double ** atom_coord;
-    std::complex<double> *eigts1_in, *eigts2_in, *eigts3_in;
-};
-
-TEST_F(Test_PAW_Cell_k, test_paw)
-{
-    // The subroutines are executed but the results are not checked yet;
-    // it is because the results are only intermediate and are private members of the class
-    // I will check the final results later once I finish them
-
-    atom_coord = new double * [2];
-    atom_coord[0] = new double [3];
-    atom_coord[0][0] = 0.0; atom_coord[0][1] = 0.0; atom_coord[0][2] = 0.0;
-    atom_coord[1] = new double [3];
-    atom_coord[1][0] = 0.25; atom_coord[1][1] = 0.25; atom_coord[1][2] = 0.25;
-
-    eigts1_in = new std::complex<double> [nat * (2 * nx + 1)];
-    eigts2_in = new std::complex<double> [nat * (2 * ny + 1)];
-    eigts3_in = new std::complex<double> [nat * (2 * nz + 1)];
-
-    filename_list.resize(1);
-    filename_list[0] = "Si_test.xml";
-
-    std::ifstream ifs_eigts("eigts.dat");
-
-    for(int i = 0; i < nat*(2*nx+1); i ++)
-    {
-        ifs_eigts >> eigts1_in[i];
-    }
-    for(int i = 0; i < nat*(2*ny+1); i ++)
-    {
-        ifs_eigts >> eigts2_in[i];
-    }
-    for(int i = 0; i < nat*(2*nz+1); i ++)
-    {
-        ifs_eigts >> eigts3_in[i];
-    }
-
-    paw_cell.init_paw_cell(ecut, cell_factor, omega, nat, ntyp,
-        atom_type, (const double **) atom_coord, filename_list);
-    paw_cell.set_eigts(nx,ny,nz,eigts1_in, eigts2_in, eigts3_in);
-
-    for(int ia = 0; ia < nat; ia ++)
-    {
-        delete[] atom_coord[ia];
-    }
-    delete[] atom_coord;
-
-    delete[] eigts1_in;
-    delete[] eigts2_in;
-    delete[] eigts3_in;
-
-    //=========================================
-
-    int npw = 411;
-    int *ig_to_ix, *ig_to_iy, *ig_to_iz;
-    double ** kpg = new double * [npw];    
-    double kpt[3] = {0.0,0.0,0.0};
-
-    ig_to_ix = new int[npw];
-    ig_to_iy = new int[npw];
-    ig_to_iz = new int[npw];
-
-    std::ifstream ifs_igxyz("igxyz.dat");
-    std::ifstream ifs_kpg("kpg1.dat");
-
-    int ig;
-    for(int i = 0; i < npw; i++)
-    {
-        ifs_igxyz >> ig >> ig_to_ix[i] >> ig_to_iy[i] >> ig_to_iz[i];
-        kpg[i] = new double[3];
-        ifs_kpg >> ig >> kpg[i][0] >> kpg[i][1] >> kpg[i][2];
-    }
-
-    int* isk_in = new int[1];
-    isk_in[0] = 0;
-    paw_cell.set_isk(1,isk_in);
-    paw_cell.set_currentk(0);
-    delete[] isk_in;
-    paw_cell.set_paw_k(npw, kpt, ig_to_ix, ig_to_iy, ig_to_iz, (const double **) kpg, tpiba);
-
-    delete[] ig_to_ix;
-    delete[] ig_to_iy;
-    delete[] ig_to_iz;
-    for(int i = 0; i < npw; i++)
-    {
-        delete[] kpg[i];
-    }
-    delete[] kpg;
-
-    paw_cell.get_vkb();
-    auto vkb = paw_cell.output_vkb();
-
-    EXPECT_EQ(vkb.size(),16);
-    EXPECT_EQ(vkb[0].size(),npw);
-
-    std::ifstream ifs_vkb("vkb_ref.dat");
-    for(int iproj = 0; iproj < 16; iproj ++)
-    {
-        for(int ipw = 0; ipw < npw; ipw ++)
-        {
-            std::complex<double> tmp;
-            ifs_vkb >> tmp;
-            EXPECT_NEAR(tmp.real(),vkb[iproj][ipw].real(),1e-8);
-            EXPECT_NEAR(tmp.imag(),vkb[iproj][ipw].imag(),1e-8);
-        }
-    }
-
-    std::complex<double> *psi;
-    psi = new std::complex<double>[npw];
-    const int nband = 6;
-    std::vector<double> weight={2,2,2,2,0,0};
-
-    std::ifstream ifs_psi("psi.dat");
-
-    paw_cell.reset_rhoij();
-    for(int iband = 0; iband < nband; iband ++)
-    {
-        for(int ipw = 0; ipw < npw; ipw ++)
-        {
-            ifs_psi >> psi[ipw];
-        }
-        paw_cell.accumulate_rhoij(psi,weight[iband]);
-    }
-
-    delete[] psi;
-
-    std::ifstream ifs_rhoij("rhoij1.dat");
-
-    std::vector<std::vector<double>> rhoijp;
-    std::vector<std::vector<int>> rhoijselect;
-    std::vector<int> nrhoijsel;
-
-    paw_cell.get_rhoijp(rhoijp, rhoijselect, nrhoijsel);
-
-    EXPECT_EQ(rhoijp.size(),nat);
-    EXPECT_EQ(rhoijselect.size(),nat);
-    EXPECT_EQ(nrhoijsel.size(),nat);
-
-    for(int iat = 0; iat < nat; iat ++)
-    {
-        // As all entries are larger than 1e-10, rhoijp is the same as rhoij
-        EXPECT_EQ(rhoijp[iat].size(),36);
-        EXPECT_EQ(rhoijselect[iat].size(),36);
-        EXPECT_EQ(nrhoijsel[iat],36);
-        for(int i = 0; i < 36; i ++)
-        {
-            double tmp;
-            ifs_rhoij >> tmp;
-            EXPECT_EQ(rhoijselect[iat][i],i+1);
-            EXPECT_NEAR(rhoijp[iat][i],tmp,1e-8);
-        }
-    }
-
-    const int nproj = 8;
-    double** dij;
-    dij = new double*[1];
-    dij[0]=new double[nproj * nproj];
-    for(int i = 0; i < nproj * nproj; i ++)
-    {
-        dij[0][i] = 0.0;
-    }
-    dij[0][0] = 13.407893;
-    dij[0][9] = 0.8201733412;
-    dij[0][18] = 5.491609854;
-    dij[0][27] = 5.491609854;
-    dij[0][36] = 5.491609854;
-    dij[0][45] = 0.59649632;
-    dij[0][54] = 0.59649632;
-    dij[0][63] = 0.59649632;
-
-    std::vector<double> sij;
-    sij.resize(nproj * nproj);
-    for(int i = 0; i < sij.size(); i ++)
-    {
-        sij[i] = 0.0;
-    }
-    sij[0] = -4.902127221589223E-002;
-    sij[9] = -9.18672607663861;
-    sij[18] = -6.319002149104143E-003;
-    sij[27] = -6.319002149104143E-003;
-    sij[36] = -6.319002149104143E-003;
-    sij[45] = -2.38515151080165;
-    sij[54] = -2.38515151080165;
-    sij[63] = -2.38515151080165;
-    
-    sij[1] = 0.726604973599628;
-    sij[8] = 0.726604973599628;
-    sij[21] = 0.156822922280989;
-    sij[30] = 0.156822922280989;
-    sij[39] = 0.156822922280989;
-    sij[42] = 0.156822922280989;
-    sij[51] = 0.156822922280989;
-    sij[60] = 0.156822922280989;
-
-    for(int iat = 0; iat < nat; iat ++)
-    {
-        paw_cell.set_dij(iat,dij);
-        paw_cell.set_sij(iat,sij.data());
-    }
-    delete[] dij[0];
-    delete[] dij;
-
-    psi = new std::complex<double>[npw];
-    std::complex<double> *vnlpsi, *snlpsi;
-    vnlpsi = new std::complex<double>[npw];
-    snlpsi = new std::complex<double>[npw];
-
-    ifs_psi.clear();
-    ifs_psi.seekg (0, std::ios::beg);
-
-    std::ifstream ifs_vnlpsi("vnlpsi_ref.dat");
-    std::ifstream ifs_snlpsi("snlpsi_ref.dat");
-
-    std::cout << std::setprecision(10);
-    for(int iband = 0; iband < nband; iband ++)
-    {
-        for(int ipw = 0; ipw < npw; ipw ++)
-        {
-            ifs_psi >> psi[ipw];
-            vnlpsi[ipw] = 0.0;
-        }
-
-        paw_cell.paw_nl_psi(0, psi, vnlpsi);
-        paw_cell.paw_nl_psi(1, psi, snlpsi);
-
-        for(int ipw = 0; ipw < npw; ipw ++)
-        {
-            std::complex<double> tmp;
-            ifs_vnlpsi >> tmp;
-            EXPECT_NEAR(tmp.real(),vnlpsi[ipw].real(),1e-8);
-            EXPECT_NEAR(tmp.imag(),vnlpsi[ipw].imag(),1e-8);
-
-            ifs_snlpsi >> tmp;
-            EXPECT_NEAR(tmp.real(),snlpsi[ipw].real(),1e-8);
-            EXPECT_NEAR(tmp.imag(),snlpsi[ipw].imag(),1e-8);
-        }
-    }
-
-    delete[] psi;
-    delete[] vnlpsi;
-    delete[] snlpsi;
-
-}
